+++ src/ops.rs
@@ -55,7 +55,12 @@ pub fn do_build(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
                 if queried_path.exists() {
                     trace!("Found {{{}}}", queried_path.display());
                     if queried_path.is_file() {
-                        trace!("{} is a file, overriding it", queried_path.display());
+                        trace!("{} is a file", queried_path.display());
+                        if ! args.force {
+                            info!("{{{}}} is ready at {{{}}}.", query, queried_path.display());
+                            info!("Try running with --force to force build.");
+                            return Ok("File was ready".to_string());
+                        }
                     } else {
                         error!("{} is not a file", queried_path.display());
                         return Err("Not a file".to_string())
@@ -88,6 +93,7 @@ pub fn do_build(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
                                  *   .expect("failed to execute process")
                                  */
                             } else {
+                                debug!("Running \'aclocal ; autoconf; automake --add-missing ; ./configure \"{}\"\'", env.configure_arg);
                                 let output = Command::new("sh")
                                     .arg("-c")
                                     .arg(format!("aclocal ; autoconf ; automake --add-missing ; ./configure \"{}\"", env.configure_arg))
@@ -178,11 +184,23 @@ pub fn do_build(env: &AmbosoEnv, args: &Args) -> Result<String,String> {
                                             Some(gsinit_ec) => {
                                                 if gsinit_ec == 0 {
                                                     debug!("Submodule init succeded with status: {}", gsinit_ec.to_string());
-                                                    let output = Command::new("sh")
-                                                        .arg("-c")
-                                                        .arg(format!("make >&2"))
-                                                        .output()
-                                                        .expect("failed to execute process");
+                                                    let output;
+                                                    if args.no_rebuild {
+                                                        debug!("Running \'make\'");
+                                                        output = Command::new("sh")
+                                                            .arg("-c")
+                                                            .arg(format!("make"))
+                                                            .output()
+                                                            .expect("failed to execute process");
+                                                    }
+                                                    else {
+                                                        debug!("Running \'make rebuild\'");
+                                                        output = Command::new("sh")
+                                                            .arg("-c")
+                                                            .arg(format!("make rebuild"))
+                                                            .output()
+                                                            .expect("failed to execute process");
+                                                    }
                                                     match output.status.code() {
                                                         Some(make_ec) => {
                                                             if make_ec == 0 {